+++ app/Services/User/statistics_service.py
@@ -1,0 +1,38 @@
+from typing import Dict, Any
+import requests
+
+from Settings.logging_config import setup_logging
+from Config.constants import LEETCODE_API
+from Config.queries import lc_query
+
+logger = setup_logging()
+
+
+class StatisticsService:
+
+    @staticmethod
+    def getAllLeetcodeData(userName: str) -> Dict[str, Any]:        
+        try:
+            response = requests.post(
+                LEETCODE_API,
+                json={"query": lc_query, "variables": {"username": userName}},
+                timeout=20
+                )
+            data = response.json()
+
+            if "errors" in data:
+                return {"leetcode": {"error": data["errors"]}}
+
+            result_data = data.get("data", {})
+
+            return {
+                "leetcode": {
+                    "profile": result_data.get("matchedUser"),
+                    "contestRanking": result_data.get("userContestRanking"),
+                    # "contestHistory": result_data.get("userContestRankingHistory")
+                }
+            }
+
+        except Exception as e:
+            logger.exception("LeetCode API fetch failed")
+            return {"error": str(e)}
